rational: specify invariants of ratio
Lay down the law on negative denominators, as (a) not all methods
accounted for them properly to begin with, and (b) those that did
can now be simplified.  In particular:

* `#[derive(PartialEq,Eq)]` instead of delegating to `Ord`.
* Methods of `Signed` mostly just delegate to the numerator now.

No outwardly visible behavior should notably change as a result of
these updates, except for people who were using `new_raw` for
nefarious purposes. (but don't take my word for it; please do check!)

This also adds a specification that the denominator of zero is 1.
(really, this is part of the same rule by which `numer` and `denom`
must share no factors in common; stated generally, the *actual*
rule is that "numer and denom must be coprime")

---

Note: I was going to stick a `debug_assert_eq!(&ret, &ret.reduced())`
in `new_raw`, but didn't after realizing I would have to add a
`T: fmt::Debug` bound. :/